pyproject.toml
[project]
name = "vopscli"
version = "0.1"
description = "Vault Operations CLI"
readme = "README.md"
authors = [{name = "Alain Chiasson"}]
license = {text = "Apache-2.0"}
requires-python = ">=3.8"
classifiers = [
    "License :: OSI Approved :: Apache Software License"
]
dependencies = [
    "click",
    "pyyaml",
    "hvac[parser]"
]

[project.urls]
Homepage = "https://github.com/alainchiasson/vopscli"
Changelog = "https://github.com/alainchiasson/vopscli/releases"
Issues = "https://github.com/alainchiasson/vopscli/issues"
CI = "https://github.com/alainchiasson/vopscli/actions"

[tool.setuptools]
packages = ["vopscli", "managedvault"]

[project.entry-points.console_scripts]
vopscli = "vopscli.cli:cli"

[project.optional-dependencies]
test = ["pytest"]
